+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+- Use ssl socket factory from custom ssl context ([#20](https://github.com/scm-manager/scm-mail-plugin/pull/20))
+
 ## 2.6.0 - 2021-06-04
 ### Added
 - Notifications for health checks ([#18](https://github.com/scm-manager/scm-mail-plugin/pull/18))